Refactor AuthServiceClient package structure and update password reset link for technicians
